TimeEval-GUI/requirements.txt
streamlit==1.11.1
timeeval==1.2.4
timeeval-gutentag==0.2.0
pyyaml
numpy
pandas
matplotlib
requests
protobuf>=3.20,<4
watchdog==2.1.9
plotly==5.10.*
